got the forms with drop down select working, onto devices and service_records next

+++ priv/repo/migrations/20240523034220_create_contracts.exs
@@ -11,7 +11,7 @@ defmodule Crm.Repo.Migrations.CreateContracts do
       add :type, :string
       add :svcalloperdevice, :integer
       add :timeallo, :integer
-      add :value, :float
+      add :value, :decimal
       add :active, :boolean, default: false, null: false
       add :company_id, references(:companies, on_delete: :nothing)
